pyproject.toml
[project]
name = "sumuppdf"
version = "0.1.0"
description = "Project to summarize and read documents using AI"
authors = [
    {name = "ramonlobo633"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "paddleocr (>=3.3.2,<4.0.0)",
    "transformers (>=4.57.5,<5.0.0)",
    "python-multipart (>=0.0.21,<0.0.22)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pillow (>=12.1.0,<13.0.0)",
    "pdf2image (>=1.17.0,<2.0.0)",
    "opencv-python (>=4.12.0.88,<5.0.0.0)",
    "paddlepaddle (>=3.3.0,<4.0.0)",
    "scikit-learn (>=1.8.0,<2.0.0)",
    "joblib (>=1.5.3,<2.0.0)"
]

[tool.poetry.dependencies]
python = ">=3.11,<3.15"
torch = "^2.9.1"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
pytest-asyncio = "^1.3.0"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "sumuppdf"
version = "0.1.0"
package-mode = false
